Update requirements.txt and scripts for TensorFlow training
