+++ requirements.txt
@@ -17,3 +17,4 @@ transformers
 seaborn
 tfswin
 tensorflow
+tensorflow-addons
