+++ src/utils/set_up.py
@@ -1,44 +1,3 @@
-import pyfiglet
-import logging
-import os
-from datetime import datetime
-from utils.create_dir import create_directory_if_not_exists
-from utils.check_gpu import check_gpu
-from datetime import datetime, timedelta
-import time
-
-def set_up_tf(args):
-    result = pyfiglet.figlet_format("Forprosjekt B)", font = "slant")  
-    print(result) 
-    project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
-
-    job_name = f'{args.date}-{args.model}-{args.class_imbalance}'
-
-    LOG_DIR = f'/cluster/home/taheeraa/code/forprosjekt/output/tf/{job_name}'
-    create_directory_if_not_exists(LOG_DIR)
-    
-    LOG_FILE = f"{LOG_DIR}/log_file.txt"
-
-    logging.basicConfig(level=logging.INFO, 
-                format='[%(levelname)s] %(asctime)s - %(message)s',
-                handlers=[
-                    logging.FileHandler(LOG_FILE),
-                    logging.StreamHandler()
-                ])
-    logger = logging.getLogger()
-    logger.info(f'Running: {args.model}')
-    logger.info(f'Root directory of project: {project_root}')
-    logger.info(f'Log directory: {LOG_DIR}')
-
-    logger.info(f'Batch size: {args.BATCH_SIZE}')
-    logger.info(f'Shuffle buffer size: {args.SHUFFLE_BUFFER_SIZE}')
-    logger.info(f'Epochs: {args.EPOCHS}')
-
-
-    check_gpu(logger)
-    logger.info('Set-up completed')
-
-    return logger, LOG_DIR
 
 def set_up(args):
     start_time = time.time()
